feat: ConnectionMonitor as layer on top of muon::Client
This change turns the `Session` itself into a `muon::Sender`.
This allows us to send requests directly via the `Session`:
`self.session().send(GET!("/tests/ping"))`.

The benefit of this is that we can wrap the entire muon client's
send implementation (including its annoying timeout behavior) with
the connection monitor layer.

Furthermore, all the API traits are now implemented for any
type `T: Sender<HttpReq, HttpRes>`, which means they apply
to the `Session` and to anything else that can send requests.

Lastly, I tried to prevent the muon client from "leaking" outside
the `Session`, so that we're sure that every network request truly
does go via the `Session`.

+++ core/core-common/src/user_context/nuke_utils.rs
@@ -147,18 +147,18 @@ async fn remove_files(paths: &[PathBuf]) -> Vec<PathBuf> {
     let mut failed = vec![];
 
     for file in paths {
-        if file.exists() {
-            if let Err(e) = fs::remove_file(file).await {
-                tracing::error!("Could not remove `{}`, details: `{e}`", file.display());
-                let ext = file
-                    .extension()
-                    .and_then(|ext| ext.to_str())
-                    .map_or("nuked".to_string(), |ext| format!("{ext}.nuked"));
-                let new_path = file.with_extension(ext);
-                match fs::rename(file, &new_path).await {
-                    Ok(()) => failed.push(new_path),
-                    Err(_) => failed.push(file.clone()),
-                }
+        if file.exists()
+            && let Err(e) = fs::remove_file(file).await
+        {
+            tracing::error!("Could not remove `{}`, details: `{e}`", file.display());
+            let ext = file
+                .extension()
+                .and_then(|ext| ext.to_str())
+                .map_or("nuked".to_string(), |ext| format!("{ext}.nuked"));
+            let new_path = file.with_extension(ext);
+            match fs::rename(file, &new_path).await {
+                Ok(()) => failed.push(new_path),
+                Err(_) => failed.push(file.clone()),
             }
         }
     }
